Add files via upload
Purpose: The purpose of this project was to create an Excel dashboard that would allow users to analyze data about bike buyers.
Data: The data used for this project was a dataset of bike buyers. The dataset included information about the buyers' age, gender, location, bike type, and purchase price.
Tools: The tools used for this project were Excel, IF functions, REPLACE, Pivot Tables, Graphs, Slicers, and Dashboards.
Process: The process for this project was as follows:
The data was cleaned using IF functions and REPLACE.
Pivot Tables were created to summarize the data.
Graphs were created to visualize the data.
Slicers were added to the graphs to allow users to filter the data.
A dashboard was created to present the data in a user-friendly way.
Results: The results of this project were a well-designed and easy-to-use Excel dashboard that allowed users to analyze data about bike buyers.